[tiny] split file per component
