+++ packages/gasket-intl/src/index.js
@@ -1,1 +1,5 @@
-export { withIntlProvider, withLocaleRequired, LocaleRequired } from './components';
+import withIntlProvider from './with-intl-provider';
+import withLocaleRequired from './with-locale-required';
+import LocaleRequired from './locale-required';
+
+export { withIntlProvider, withLocaleRequired, LocaleRequired };
